+++ dashboard/templates/wagtailadmin/home.html
@@ -2,5 +2,6 @@
 {% load i18n %}
 
 {% block branding_welcome %}
-  {% translate "Welcome to the administration panel of Sites faciles" %}
+  {% translate "Sites faciles" %}
 {% endblock %}
+
